+++ flink-connector/src/main/java/org/apache/gravitino/flink/connector/hive/GravitinoHiveCatalog.java
@@ -24,9 +24,9 @@ import org.apache.flink.table.catalog.AbstractCatalog;
 import org.apache.flink.table.catalog.exceptions.CatalogException;
 import org.apache.flink.table.catalog.hive.HiveCatalog;
 import org.apache.flink.table.factories.Factory;
-import org.apache.gravitino.flink.connector.DefaultTransformConverter;
+import org.apache.gravitino.flink.connector.DefaultPartitionConverter;
+import org.apache.gravitino.flink.connector.PartitionConverter;
 import org.apache.gravitino.flink.connector.PropertiesConverter;
-import org.apache.gravitino.flink.connector.TransformConverter;
 import org.apache.gravitino.flink.connector.catalog.BaseCatalog;
 import org.apache.hadoop.hive.conf.HiveConf;
 
@@ -74,8 +74,8 @@ public class GravitinoHiveCatalog extends BaseCatalog {
   }
 
   @Override
-  protected TransformConverter getTransformConverter() {
-    return DefaultTransformConverter.INSTANCE;
+  protected PartitionConverter getPartitionConverter() {
+    return DefaultPartitionConverter.INSTANCE;
   }
 
   @Override
